refactor search bar
